support for non-integer percentiles (trac https://trac.osgeo.org/grass/ticket/873; merge from devbr6)
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@40966 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ raster/r.univar/globals.h
@@ -31,7 +31,7 @@ typedef struct
     double min;
     double max;
     unsigned int n_perc;
-    int *perc;
+    double *perc;
     double sum_abs;
     int n;
     int size;
